Update CUD Video

+++ Services/Fsel.Course/Fsel.Course.Domain/Entities/VideoTimeCode.cs
@@ -13,10 +13,10 @@ namespace Fsel.Course.Domain.Entities
 {
     public class VideoTimeCode : Entity
     {
-        [Range(1, int.MaxValue, ErrorMessage = nameof(EnumVideoTimeCodeErrorCode.VTC04C))]
+        [Range(1, long.MaxValue, ErrorMessage = nameof(EnumVideoTimeCodeErrorCode.VTC04C))]
         public long DisplayTimeTicks { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = nameof(EnumVideoTimeCodeErrorCode.VTC04C))]
+        [Range(1, long.MaxValue, ErrorMessage = nameof(EnumVideoTimeCodeErrorCode.VTC04C))]
         public long ExecutionTimeTicks { get; set; }
 
         public EnumTimeCodeType TimeCodeType { get; set; }
